Fix review comments and refactor the code.
