+++ apps/myaccount/src/components/shared/system-notification-alert.tsx
@@ -17,22 +17,11 @@
  */
 
 import Alert from "@oxygen-ui/react/Alert";
-import { ProfileConstants } from "@wso2is/core/constants";
 import { IdentifiableComponentInterface } from "@wso2is/core/models";
-import {  LinkButton } from "@wso2is/react-components";
-import { AxiosError } from "axios";
-import React, { FunctionComponent, ReactElement, useEffect, useState } from "react";
-import { useTranslation } from "react-i18next";
-import { useDispatch, useSelector } from "react-redux";
-import { Dispatch } from "redux";
-import { getPreference, resendVerificationLinkOrCode } from "../../api";
-import { ProfileConstants as MyAccountProfileConstants, RecoveryScenario } from "../../constants";
-import { SCIMConfigs } from "../../extensions/configs/scim";
-import { AlertLevels, AuthStateInterface } from "../../models";
-import { PreferenceConnectorResponse, PreferenceProperty, PreferenceRequest } from "../../models/preference";
-import { AppState } from "../../store";
-import { addAlert } from "../../store/actions";
+import React, { FunctionComponent, ReactElement } from "react";
 import "./system-notification-alert.scss";
+import { useResendAccountConfirmationAlert } from "../../hooks/use-resend-account-confirmation-alert";
+import { SystemNotificationAlertState } from "../../models";
 
 /**
  * SystemNotificationAlert component props.
@@ -45,163 +34,21 @@ export type SystemNotificationAlertProps = IdentifiableComponentInterface;
  * @param props - Props injected into the component.
  * @returns ReactElement
  */
-export const SystemNotificationAlert: FunctionComponent<SystemNotificationAlertProps> = (
-    props: SystemNotificationAlertProps
-): ReactElement | null => {
-    const { [ "data-componentid" ]: componentId } = props;
+export const SystemNotificationAlert: FunctionComponent<SystemNotificationAlertProps> = ({
+    ["data-componentid"]: componentId = "system-notification-alert"
+}: SystemNotificationAlertProps): ReactElement | null => {
+    const systemNotificationAlertState: SystemNotificationAlertState =
+        useResendAccountConfirmationAlert(componentId);
 
-    const { t } = useTranslation();
-    const dispatch: Dispatch = useDispatch();
+    if (!systemNotificationAlertState?.isVisible) return null;
 
-    const profileDetails: AuthStateInterface = useSelector((state: AppState) => state.authenticationInformation);
-
-    const [ isSRSendOtpInEmailEnabled, setIsSRSendOtpInEmailEnabled ] = useState<boolean>(false);
-    const [ isAccountStatePendingSR, setIsAccountStatePendingSR ] = useState<boolean>(false);
-    const [ isPreferredChannelEmail, setIsPreferredChannelEmail ] = useState<boolean>(false);
-    const [ isSubmitting, setIsSubmitting ] = useState<boolean>(false);
-
-    useEffect(() => {
-        getSelfSignUpPreferences();
-    }, []);
-
-    useEffect(() => {
-        const accountState: string = profileDetails
-            ?.profileInfo
-            ?.[SCIMConfigs?.scim?.systemSchema]
-            ?.[ProfileConstants.SCIM2_SCHEMA_DICTIONARY.get(MyAccountProfileConstants.ACCOUNT_STATE)];
-
-        const preferredChannel: string = profileDetails
-            ?.profileInfo
-            ?.[SCIMConfigs?.scim?.systemSchema]
-            ?.[ProfileConstants.SCIM2_SCHEMA_DICTIONARY.get(MyAccountProfileConstants.PREFERRED_CHANNEL)];
-
-        setIsAccountStatePendingSR(accountState === MyAccountProfileConstants.ACCOUNT_STATE_PENDING_SR);
-        setIsPreferredChannelEmail(preferredChannel === MyAccountProfileConstants.PREFERRED_CHANNEL_EMAIL);
-    }, [ profileDetails?.profileInfo ]);
-
-    /**
-     * Gets the self sign up account confirmation preferences.
-     */
-    const getSelfSignUpPreferences = (): void => {
-        const selfSignUpConnector: PreferenceRequest[] = [ {
-            "connector-name": ProfileConstants.SELF_SIGN_UP_CONNECTOR,
-            properties: [ ProfileConstants.SELF_SIGN_UP_ENABLE_SEND_OTP_IN_EMAIL ]
-        } ];
-
-        getPreference(selfSignUpConnector)
-            .then((response: PreferenceConnectorResponse[]) => {
-                if (response) {
-                    const selfSignUpOptions: PreferenceConnectorResponse[] = response;
-                    const responseProperties: PreferenceProperty[] = selfSignUpOptions[0].properties;
-
-                    const sendOtpInEmailProperty: PreferenceProperty = responseProperties?.find(
-                        (prop: PreferenceProperty) =>
-                            prop.name === ProfileConstants.SELF_SIGN_UP_ENABLE_SEND_OTP_IN_EMAIL
-                    );
-                    const isSendOtpInEmailEnabled: boolean = sendOtpInEmailProperty?.value === "true";
-
-                    setIsSRSendOtpInEmailEnabled(isSendOtpInEmailEnabled);
-                } else {
-                    dispatch(addAlert({
-                        description: t(
-                            "myAccount:components.selfSignUp.preference.notifications.genericError.description"
-                        ),
-                        level: AlertLevels.ERROR,
-                        message: t(
-                            "myAccount:components.selfSignUp.preference.notifications.genericError.message"
-                        )
-                    }));
-                }
-            })
-            .catch((error: AxiosError) => {
-                if (error?.response?.data?.detail) {
-                    dispatch(addAlert({
-                        description: t(
-                            "myAccount:components.selfSignUp.preference.notifications.error.description",
-                            { description: error.response.data.detail }
-                        ),
-                        level: AlertLevels.ERROR,
-                        message: t("myAccount:components.selfSignUp.preference.notifications.error.message")
-                    }));
-
-                    return;
-                }
-
-                dispatch(addAlert({
-                    description: t(
-                        "myAccount:components.selfSignUp.preference.notifications.genericError.description"
-                    ),
-                    level: AlertLevels.ERROR,
-                    message: t("myAccount:components.selfSignUp.preference.notifications.genericError.message")
-                }));
-            });
-    };
-
-    /**
-     * Resends the account confirmation Email link.
-     */
-    const resendAccountConfirmationLink = (): void => {
-        if (isSubmitting) {
-            return;
-        }
-        setIsSubmitting(true);
-        resendVerificationLinkOrCode(RecoveryScenario.SELF_SIGN_UP)
-            .then(() => {
-                dispatch(addAlert({
-                    description: t("myAccount:components.systemNotificationAlert.selfSignUp." +
-                        "notifications.resendSuccess.description"),
-                    level: AlertLevels.SUCCESS,
-                    message: t("myAccount:components.systemNotificationAlert.selfSignUp." +
-                        "notifications.resendSuccess.message")
-                }));
-            })
-            .catch((errorMessage: AxiosError) => {
-                dispatch(addAlert({
-                    description: t("myAccount:components.systemNotificationAlert.selfSignUp." +
-                        "notifications.resendError.description", { error: errorMessage }),
-                    level: AlertLevels.ERROR,
-                    message: t("myAccount:components.systemNotificationAlert.selfSignUp." +
-                        "notifications.resendError.message")
-                }));
-            }).finally(() => {
-                setIsSubmitting(false);
-            });
-    };
-
-    /*
-    * The resending capability is available only when account confirmation is configured
-    * to use an Email Link. The rationale for excluding Email OTP and SMS OTP confirmation
-    *  methods is that their respective code validations must occur during the initial
-    * user registration flow itself rather in a later stage via MyAccount.
-    */
-    if (isAccountStatePendingSR && isPreferredChannelEmail && !isSRSendOtpInEmailEnabled) {
-        return (
-            <Alert
-                severity={ AlertLevels.WARNING }
-                data-componentid={ componentId }
-                className="system-notification-alert"
-            >
-                { t("myAccount:components.systemNotificationAlert.selfSignUp." +
-                        "awaitingAccountConfirmation") }
-                <LinkButton
-                    onClick={ resendAccountConfirmationLink }
-                    aria-disabled={ isSubmitting }
-                    disabled={ isSubmitting }
-                    data-testid={ `${ componentId }-resend-link` }
-                >
-                    { t("myAccount:components.systemNotificationAlert.resend") }
-                </LinkButton>
-            </Alert>
-        );
-    }
-
-    return null;
+    return (
+        <Alert
+            severity={ systemNotificationAlertState?.severity }
+            data-componentid={ componentId }
+            className="system-notification-alert"
+        >
+            { systemNotificationAlertState?.message }
+        </Alert>
+    );
 };
-
-/**
- * Default properties of {@link SystemNotificationAlert}
- */
-SystemNotificationAlert.defaultProps = {
-    "data-componentid": "system-notification-alert"
-};
-
