+++ apps/myaccount/src/constants/profile-constants.ts
@@ -58,6 +58,7 @@ export class ProfileConstants {
      */
     public static readonly ACCOUNT_STATE: string = "ACCOUNT_STATE";
     public static readonly PREFERRED_CHANNEL: string = "PREFERRED_CHANNEL";
+    public static readonly EMAIL_VERIFIED: string = "EMAIL_VERIFIED";
 }
 
 /**
